+++ the_gatehouse/urls.py
@@ -10,8 +10,8 @@ from .views import (user_settings, player_page_view,
                     guild_join_request, guild_invite_view, pending_guild_invites, pending_posts,
                     approve_guild_invite, reject_guild_invite, mark_guild_invite_clicked,
                     approve_post, reject_post, dismiss_notification,
-                    survey_list_view, survey_detail_view, survey_results_view, my_surveys_view,
-                    create_survey_view, edit_survey_view, preview_survey_view, duplicate_survey_view)
+                    survey_list_view, survey_results_view, my_surveys_view, survey_take_view, survey_response_view, survey_response_edit_view, survey_redirect, survey_detail_view,
+                    create_survey_view, survey_edit_view, survey_preview_view, duplicate_survey_view)
 from the_warroom.views import player_game_list_view  #PlayerGameListView
 urlpatterns = [
     # path("", list_view, name='home'),
@@ -75,13 +75,20 @@ urlpatterns = [
 
     # Survey URLs
     path('surveys/', survey_list_view, name='survey-list'),
-    path('surveys/create/', create_survey_view, name='create-survey'),
+    path('surveys/create/', create_survey_view, name='survey-create'),
+
     path('surveys/<slug:slug>/', survey_detail_view, name='survey-detail'),
-    path('surveys/<slug:slug>/edit/', edit_survey_view, name='edit-survey'),
-    path('surveys/<slug:slug>/preview/', preview_survey_view, name='preview-survey'),
-    path('surveys/<slug:slug>/duplicate/', duplicate_survey_view, name='duplicate-survey'),
+    path('surveys/<slug:slug>/redirect/', survey_redirect, name='survey-redirect'),
+    path('surveys/<slug:slug>/take/', survey_take_view, name='survey-take'),
+    path('surveys/<slug:slug>/edit/', survey_edit_view, name='survey-edit'),
+    path('surveys/<slug:slug>/preview/', survey_preview_view, name='survey-preview'),
+    path('surveys/<slug:slug>/duplicate/', duplicate_survey_view, name='survey-duplicate'),
     path('surveys/<slug:slug>/results/', survey_results_view, name='survey-results'),
     path('profile/<slug:slug>/surveys/', my_surveys_view, name='my-surveys'),
+
+    path('surveys/<slug:slug>/responses/<int:response_id>/', survey_response_view, name='survey-user-response'),
+    path('surveys/<slug:slug>/responses/<int:response_id>/edit/', survey_response_edit_view, name='survey-edit-response'),
+
 
     # path('fake-error/', trigger_error),
     # path('test-error/', trigger_other_error),